services/SignSrv/MediaCrawlerPro-SignSrv/pyproject.toml
[project]
name = "mediacrawlerpro-signsrv"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx==0.27.0",
    "playwright==1.42.0",
    "pydantic==2.8.2",
    "pyexecjs==1.5.1",
    "tenacity==9.0.0",
    "tornado==6.4",
    "pre-commit>=3.5.0",
]

[[tool.uv.index]]
url = "https://mirrors.aliyun.com/pypi/simple/"
default = true

